Add BaseDatabase
