+++ spinn_front_end_common/interface/buffer_management/storage_objects/buffer_database.py
@@ -16,9 +16,8 @@
 import os
 import sqlite3
 import time
-from spinn_utilities.abstract_context_manager import AbstractContextManager
 from spinn_front_end_common.data import FecDataView
-from spinn_front_end_common.utilities.sqlite_db import SQLiteDB
+from .base_database import BaseDatabase
 
 _DDL_FILE = os.path.join(os.path.dirname(__file__), "db.sql")
 _SECONDS_TO_MICRO_SECONDS_CONVERSION = 1000
@@ -29,7 +28,7 @@ def _timestamp():
     return int(time.time() * _SECONDS_TO_MICRO_SECONDS_CONVERSION)
 
 
-class BufferDatabase(SQLiteDB, AbstractContextManager):
+class BufferDatabase(BaseDatabase):
     """ Specific implementation of the Database for SQLite 3.
 
     There should only ever be a single Database Object in use at any time.
@@ -47,30 +46,6 @@ class BufferDatabase(SQLiteDB, AbstractContextManager):
 
     __slots__ = []
 
-    def __init__(self, database_file=None, read_only=False):
-        """
-        :param str database_file:
-            The name of a file that contains (or will contain) an SQLite
-            database holding the data.
-            If omitted the default location will be used.
-        """
-        if database_file is None:
-            database_file = self.default_database_file()
-
-        if read_only:
-            super().__init__(database_file, read_only=False)
-        else:
-            super().__init__(
-                database_file, read_only=False, ddl_file=_DDL_FILE)
-
-    @classmethod
-    def default_database_file(cls):
-        if FecDataView.get_reset_number():
-            return os.path.join(
-                FecDataView.get_run_dir_path(),
-                f"buffer{FecDataView.get_reset_number()}.sqlite3")
-        return os.path.join(
-            FecDataView.get_run_dir_path(), "buffer.sqlite3")
 
     def clear_region(self, x, y, p, region):
         """ Clears the data for a single region.
@@ -154,28 +129,6 @@ class BufferDatabase(SQLiteDB, AbstractContextManager):
                 idx += len(item)
             data = c_buffer
         return memoryview(data)
-
-    @staticmethod
-    def __get_core_id(cursor, x, y, p):
-        """
-        :param ~sqlite3.Cursor cursor:
-        :param int x:
-        :param int y:
-        :param int p:
-        :rtype: int
-        """
-        for row in cursor.execute(
-                """
-                SELECT core_id FROM region_view
-                WHERE x = ? AND y = ? AND processor = ?
-                LIMIT 1
-                """, (x, y, p)):
-            return row["core_id"]
-        cursor.execute(
-            """
-            INSERT INTO core(x, y, processor) VALUES(?, ?, ?)
-            """, (x, y, p))
-        return cursor.lastrowid
 
     def __get_region_id(self, cursor, x, y, p, region):
         """
